refactor: 유틸 함수 모듈화
